Move container customization to its own CompilerPass
The code is rather specific, so it is not justified to have it live
directly inside Kernel.php.
