+++ api/src/Kernel.php
@@ -2,34 +2,15 @@
 
 namespace App;
 
-use Doctrine\ORM\EntityManagerInterface;
-use Lexik\Bundle\JWTAuthenticationBundle\Encoder\JWTEncoderInterface;
+use App\DependencyInjection\JWTOAuthStateCompilerPass;
 use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
-use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\HttpKernel\Kernel as BaseKernel;
 use Symfony\Component\Routing\Loader\Configurator\RoutingConfigurator;
 
-class Kernel extends BaseKernel implements CompilerPassInterface {
+class Kernel extends BaseKernel {
     use MicroKernelTrait;
-
-    /**
-     * In this method, we can add our own compiler pass code, i.e. customization to bundle service configs,
-     * when this isn't possible via services.yaml.
-     */
-    public function process(ContainerBuilder $container) {
-        // Add custom arguments to each of our OAuth clients, so they don't have to use native PHP sessions.
-        // See JWTStateOAuth2Client for more info.
-        foreach (array_keys($container->getExtensionConfig('knpu_oauth2_client')[0]['clients']) as $id) {
-            $definition = $container->getDefinition('knpu.oauth2.client.'.$id);
-            $definition->addArgument('%env(COOKIE_PREFIX)%');
-            $definition->addArgument('%kernel.environment%');
-            $definition->addArgument(new Reference(JWTEncoderInterface::class));
-            $definition->addArgument(new Reference(EntityManagerInterface::class));
-        }
-    }
 
     protected function configureContainer(ContainerConfigurator $container): void {
         $container->import('../config/{packages}/*.yaml');
@@ -56,5 +37,7 @@ class Kernel extends BaseKernel implements CompilerPassInterface {
 
     protected function build(ContainerBuilder $container) {
         parent::build($container);
+
+        $container->addCompilerPass(new JWTOAuthStateCompilerPass());
     }
 }
